move utils and definitions to central place
