+++ aixweather/core_data_format_2_output_file/unconverted_to_x.py
@@ -6,7 +6,7 @@ import json
 import pickle
 import pandas as pd
 
-from aixweather.core_data_format_2_output_file import utils_2output
+from aixweather import definitions
 from aixweather.imports.utils_import import MetaData
 
 
@@ -28,8 +28,8 @@ def to_pickle(
     core_df = core_df.copy()
     filename = f"Station_{meta.station_name}.pkl"
     meta_filename = f"Station_{meta.station_name}_meta_data.pkl"
-    file_path = utils_2output.results_file_path(filename, result_folder)
-    meta_file_path = utils_2output.results_file_path(meta_filename, result_folder)
+    file_path = definitions.results_file_path(filename, result_folder)
+    meta_file_path = definitions.results_file_path(meta_filename, result_folder)
 
     core_df.to_pickle(file_path)
     with open(meta_file_path, "wb") as file:
@@ -59,8 +59,8 @@ def to_json(
 
     filename = f"Station_{meta.station_name}.json"
     meta_filename = f"Station_{meta.station_name}_meta_data.json"
-    file_path = utils_2output.results_file_path(filename, result_folder)
-    meta_file_path = utils_2output.results_file_path(meta_filename, result_folder)
+    file_path = definitions.results_file_path(filename, result_folder)
+    meta_file_path = definitions.results_file_path(meta_filename, result_folder)
 
     # Convert DataFrame to JSON and save to file
     core_df.to_json(file_path, orient="records")
@@ -94,8 +94,8 @@ def to_csv(
 
     filename = f"Station_{meta.station_name}.csv"
     meta_filename = f"Station_{meta.station_name}_meta_data.json"
-    file_path = utils_2output.results_file_path(filename, result_folder)
-    meta_file_path = utils_2output.results_file_path(meta_filename, result_folder)
+    file_path = definitions.results_file_path(filename, result_folder)
+    meta_file_path = definitions.results_file_path(meta_filename, result_folder)
 
     core_df.to_csv(file_path, index=True)
