--- config/config.ini
@@ -1,3 +1,0 @@
-[http_credentials]
-username=****@eonerc.rwth-aachen.de
-password=****
