+++ tests/test_DWD_historical.py
@@ -8,10 +8,9 @@ import datetime as dt
 
 from parameterized import parameterized
 
-from config.definitions import ROOT_DIR
+from aixweather import definitions
 from aixweather.project_class import ProjectClassDWDHistorical
 from tests import utils_4_tests
-from aixweather.core_data_format_2_output_file import utils_2output
 
 
 class BaseDWDHistorical(unittest.TestCase):
@@ -19,7 +18,7 @@ class BaseDWDHistorical(unittest.TestCase):
     def init_and_run_DWD_historical(
         cls, name: str, start: dt.datetime, end: dt.datetime, station=15000
     ):
-        abs_result_folder_path = os.path.join(utils_2output.result_folder_path(), name)
+        abs_result_folder_path = os.path.join(definitions.result_folder_path(), name)
         cls.c = ProjectClassDWDHistorical(
             start=start,
             end=end,
@@ -27,7 +26,7 @@ class BaseDWDHistorical(unittest.TestCase):
             abs_result_folder_path=abs_result_folder_path,
         )
         cls.folder_tests = os.path.join(
-            ROOT_DIR, f"tests/test_files/regular_tests/DWD_hist/test_{name}"
+            definitions.ROOT_DIR, f"tests/test_files/regular_tests/DWD_hist/test_{name}"
         )
         cls.start_formatted = start.strftime("%Y%m%d")
         cls.end_formatted = end.strftime("%Y%m%d")
