+++ tests/test_TRY.py
@@ -5,21 +5,20 @@ includes unittests for different TRY datasets
 import os
 import unittest
 
+from aixweather import definitions
 from aixweather.project_class import ProjectClassTRY
-from aixweather.core_data_format_2_output_file import utils_2output
-from config.definitions import ROOT_DIR
 from tests import utils_4_tests
 
 
 class BaseTRY(unittest.TestCase):
     @classmethod
     def init_and_run_TRY(cls, name: str, path: str):
-        abs_result_folder_path = os.path.join(utils_2output.result_folder_path(), name)
+        abs_result_folder_path = os.path.join(definitions.result_folder_path(), name)
         cls.c = ProjectClassTRY(
             path=path, abs_result_folder_path=abs_result_folder_path
         )
         cls.folder_tests = os.path.join(
-            ROOT_DIR, f"tests/test_files/regular_tests/TRY/test_{name}"
+            definitions.ROOT_DIR, f"tests/test_files/regular_tests/TRY/test_{name}"
         )
 
         utils_4_tests.run_all_functions(cls.c)
@@ -40,7 +39,7 @@ class TestDWDTRY2015(BaseTRY, utils_4_tests.RegressionTestsClass):
         cls.init_and_run_TRY(
             "TRY2015",
             os.path.join(
-                ROOT_DIR,
+                definitions.ROOT_DIR,
                 r"tests/test_files/regular_tests/TRY/"
                 "test_TRY2015/input/TRY2015_507931060546_Jahr.dat",
             ),
@@ -53,7 +52,7 @@ class TestDWDTRY2015Sommer(BaseTRY, utils_4_tests.RegressionTestsClass):
         cls.init_and_run_TRY(
             "TRY2015_Sommer",
             os.path.join(
-                ROOT_DIR,
+                definitions.ROOT_DIR,
                 r"tests/test_files/regular_tests/TRY/"
                 "test_TRY2015_Sommer/input/TRY2015_507931060546_Somm.dat",
             ),
@@ -66,7 +65,7 @@ class TestDWDTRY2045(BaseTRY, utils_4_tests.RegressionTestsClass):
         cls.init_and_run_TRY(
             "TRY2045",
             os.path.join(
-                ROOT_DIR,
+                definitions.ROOT_DIR,
                 r"tests/test_files/regular_tests/TRY/"
                 "test_TRY2045/input/TRY2045_507931060546_Jahr.dat",
             ),
