+++ tests/test_output_functions.py
@@ -9,16 +9,14 @@ import json
 
 import pandas as pd
 
+from aixweather import definitions
 from tests import utils_4_tests
-from config.definitions import ROOT_DIR
 from aixweather.imports.utils_import import MetaData
 from aixweather.project_class import ProjectClassDWDHistorical
-from aixweather.core_data_format_2_output_file import utils_2output
-
 
 class BaseOutputFunction(unittest.TestCase):
     def init(cls, name: str, start: dt.datetime, end: dt.datetime, station=15000):
-        abs_result_folder_path = os.path.join(utils_2output.result_folder_path(), name)
+        abs_result_folder_path = os.path.join(definitions.result_folder_path(), name)
         cls.c = ProjectClassDWDHistorical(
             start=start,
             end=end,
@@ -26,7 +24,7 @@ class BaseOutputFunction(unittest.TestCase):
             abs_result_folder_path=abs_result_folder_path,
         )
         cls.folder_tests = os.path.join(
-            ROOT_DIR,
+            definitions.ROOT_DIR,
             f"tests/test_files/regular_tests/output_functions/test_{name}",
         )
         cls.start_formatted = start.strftime("%Y%m%d")
